dcp-712 Fix creation of direct links from external links and reorganise tests
